Move function to utils and rename regex
