+++ app/javascript/app/components/sdg-card/sdg-card-component.jsx
@@ -2,6 +2,7 @@ import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
 import cx from 'classnames';
 import Icon from 'components/icon';
+import { compareIndex } from 'utils/utils';
 import styles from './sdg-card-styles.scss';
 
 class SDGCard extends PureComponent {
@@ -36,29 +37,6 @@ class SDGCard extends PureComponent {
 
     const title = square ? goal.title : `${goal.number}. ${goal.cw_title}`;
 
-    const compare = (a, b) => {
-      const divideRegex = /(.*)\.(.*)/;
-      const match = x => x.number.match(divideRegex);
-      const fullNumber = x =>
-        (match(x) && match(x)[1]) || x.number.match(divideRegex);
-      const decimalNumber = x => match(x) && match(x)[2];
-      const fullA = fullNumber(a);
-      const fullB = fullNumber(b);
-      const decimalA = decimalNumber(a);
-      const decimalB = decimalNumber(b);
-      if (
-        !decimalA ||
-        !decimalB ||
-        parseInt(fullA, 10) !== parseInt(fullB, 10)
-      ) {
-        return fullA > fullB ? 1 : -1;
-      }
-      if (decimalA.match(/[^0-9.]/) || decimalA.match(/[^0-9.]/)) {
-        return decimalA > decimalB ? 1 : -1;
-      }
-      return parseInt(decimalA, 10) - parseInt(decimalB, 10);
-    };
-
     return (
       <div
         className={cardStyle}
@@ -70,7 +48,7 @@ class SDGCard extends PureComponent {
         <h4 className={styles.title}>{title}</h4>
         <div className={styles.dots}>
           {targets &&
-            targets.sort(compare).map(target => {
+            targets.sort(compareIndex).map(target => {
               const isSmall =
                 target.sectors &&
                 activeSector &&
