+++ app/javascript/app/utils/utils.js
@@ -10,6 +10,26 @@ export function isCountryIncluded(countriesIncluded = [], iso) {
   return countriesIncluded.includes(iso);
 }
 
+export function compareIndex(a, b) {
+  const divideByDot = /(.*)\.(.*)/;
+  const isNotNumber = /[^0-9.]/;
+  const match = x => x.number.match(divideByDot);
+  const fullNumber = x =>
+    (match(x) && match(x)[1]) || x.number.match(divideByDot);
+  const decimalNumber = x => match(x) && match(x)[2];
+  const fullA = fullNumber(a);
+  const fullB = fullNumber(b);
+  const decimalA = decimalNumber(a);
+  const decimalB = decimalNumber(b);
+  if (!decimalA || !decimalB || parseInt(fullA, 10) !== parseInt(fullB, 10)) {
+    return fullA > fullB ? 1 : -1;
+  }
+  if (decimalA.match(isNotNumber) || decimalA.match(isNotNumber)) {
+    return decimalA > decimalB ? 1 : -1;
+  }
+  return parseInt(decimalA, 10) - parseInt(decimalB, 10);
+}
+
 export default {
   deburrUpper,
   isCountryIncluded
